refac: add modules and crate console
Modularizing the code and adding the console crate.
