+++ src/todo.rs
@@ -1,6 +1,4 @@
-pub mod todo {
-    #[derive(Debug, Clone)]
-    pub struct Todo {
-        pub message: String,
-    }
+#[derive(Debug, Clone)]
+pub struct Todo {
+    pub message: String,
 }
